+++ app/resources/resourcesCtrl.js
@@ -7,7 +7,8 @@
 
 			vm.lookupCertIds = lookupCertIds;
 			vm.download = download;
-			vm.lookupProductsFormatInvalid = false;
+			vm.lookupProductsFormatInvalidIds = [];
+			vm.lookupProductsCertIdNotFound = [];
 			vm.viewProduct = viewProduct;
 
             activate();
@@ -15,6 +16,8 @@
 			// Restore lookup IDs and results
 			vm.certIds = $localStorage.lookupCertIds;
 			vm.lookupProducts = $localStorage.lookupProducts;
+			vm.lookupProductsFormatInvalidIds = $localStorage.lookupProductsFormatInvalidIds;
+			vm.lookupProductsCertIdNotFound = $localStorage.lookupProductsCertIdNotFound;
 
 			if ($localStorage.lookupCertIds && !$localStorage.lookupProducts) {
 				lookupCertIds();
@@ -103,7 +106,8 @@
 
 			function lookupCertIds () {
 				vm.lookupProducts = null;
-				vm.lookupProductsFormatInvalid = false;
+				vm.lookupProductsFormatInvalidIds = [];
+				vm.lookupProductsCertIdNotFound = [];
 
 				if ((vm.lookup !== "undefined") && (vm.certIds !== "undefined")) {
 					vm.certIds = vm.certIds.replace(/[;,\s]+/g, " ");
@@ -111,43 +115,63 @@
 
 					// Check format of input
 					if ("" === vm.certIds.trim()) {
-						vm.lookupProductsFormatInvalid = false;
 						clearLookup();
-					} else if (null !== vm.certIds.match(/^([0-9A-Z]{15}([ ][0-9A-Z]{15})*)$/i)) {
+					} else {
 
 						// Split IDs
 						var idArray = vm.certIds.split(/ /);
 						vm.lookupProducts = null;
 
 						$localStorage.lookupCertIds = vm.certIds;
-
+						
+						var preventDuplicationIds = {};
+						
 						// Call LookupAPI
 						idArray.forEach(function (id) {
-							commonService.lookupCertificationId(id)
-								.then(function (data) {
-									if (vm.lookupProducts === null)
-										vm.lookupProducts = [];
-									data.products.forEach(function (product) {
-										product.certificationId = data.ehrCertificationId;
-										product.certificationIdEdition = data.year;
-										vm.lookupProducts.push(product);
-									});
 
-									$localStorage.lookupProducts = vm.lookupProducts;
+							// Check if we've already checked this ID in case the user entered duplicates
+							if (typeof preventDuplicationIds[id] === "undefined") {
+								preventDuplicationIds[id] = true;
+						
+								// Check if ID format is valid
+								if (!id.match(/^[0-9A-Z]{15}$/i)) {
+									// Invalid ID format
+									vm.lookupProductsFormatInvalidIds.push(id);
+								} else {
+									// Valid ID format
+									commonService.lookupCertificationId(id)
+										.then(function (data) {
 
-								}, function (error) {
-									console.debug("Error: " + error);
-									clearLookupResults();
-								});
+											if (vm.lookupProducts === null)
+												vm.lookupProducts = [];
+
+											// If the ID was found, then I have data...
+											if (data.products.length > 0) {
+												data.products.forEach(function (product) {
+													product.certificationId = data.ehrCertificationId;
+													product.certificationIdEdition = data.year;
+													vm.lookupProducts.push(product);
+												});
+											} else {
+												// ...otherwise, if the ID was not found, tell the user.
+												vm.lookupProductsCertIdNotFound.push(id);
+											}
+											
+											$localStorage.lookupProducts = vm.lookupProducts;
+											$localStorage.lookupProductsFormatInvalidIds = vm.lookupProductsFormatInvalidIds;
+											$localStorage.lookupProductsCertIdNotFound = vm.lookupProductsCertIdNotFound;
+											
+										}, function (error) {
+											console.debug("Error: " + error);
+											clearLookupResults();
+										});
+								}
+							}
 						});
-					} else {
-						$localStorage.lookupCertIds = vm.certIds;
-						vm.lookupProductsFormatInvalid = true;
-						clearLookupResults();
 					}
 				}
 			}
-
+			
 			function clearLookup() {
 				delete $localStorage.lookupCertIds;
 				vm.certIds = null;
@@ -156,6 +180,8 @@
 
 			function clearLookupResults() {
 				delete $localStorage.lookupProducts;
+				delete $localStorage.lookupProductsFormatInvalidIds
+				delete $localStorage.lookupProductsCertIdNotFound;
 				vm.lookupProducts = null;
 			}